CASMTRIAGE-3600 Use Native Python Install (#6)
* CASMTRIAGE-3600 Use SLES Native Python for RPMs

SLES15 SP2 and SP3 come with Python3.6, however our internal installs of
SP2 have an older version of glibc. Python 3.8 is being used to build
the RPM requires a glibc of 2.29 or newer, but SLES15 SP2 only has 2.26.

This conforms the RPM to build under Python3.6, which will be compatible
in SLES15 distributions (SP0, SP1, SP2, SP3, and SP4).

* Bump release for 3.5.0 patch

* Remove useless BUILD_METADATA